requirements.txt
boto3
urllib3
botocore
requests
typing-extensions
charset-normalizer
certifi
setuptools
s3transfer
wheel
packaging
idna
python-dateutil
cryptography
six
pip
numpy
pyyaml
awscli
pytz
jmespath
docutils
rsa
pyasn1-modules
pyasn1
cachetools
oauthlib
google-auth-oauthlib
matplotlib
xlwt